Separated validation, so it's easier to extend
