+++ src/components/settings/types.ts
@@ -1,4 +1,5 @@
 import { GeneratorConfig } from '../../state/types'
+import { ConfigIssues } from '../../state/validation/types'
 import { ModalPage } from '../PagedModal'
 import { SettingsPageId } from './SettingsPageId'
 
@@ -6,26 +7,6 @@ export type SelectItem<T> = {
   label: string
   value: T
 }
-
-export type Issue = {
-  type: 'error' | 'warning'
-  label: string
-}
-
-export type ConfigIssues = Partial<{
-  bars: Issue
-  bpm: Issue
-  clef: Issue
-  keySignature: Issue
-  showChordsStaff: Issue
-  showChordSymbols: Issue
-  useSeventhChords: Issue
-  notes: Issue
-  noteDurations: Issue
-  dottedNoteDurations: Issue
-  restDurations: Issue
-  dottedRestDurations: Issue
-}>
 
 export type SettingsPageProps = {
   onClose: () => void
