+++ src/state/validation/ok.ts
@@ -1,0 +1,3 @@
+export function ok(): undefined {
+  return undefined
+}
